+++ packages/bundle/test-d/fail-once/create-direct-line-invalid-option.test-d.tsx
@@ -2,7 +2,7 @@ import '../setup';
 
 import { expectNotAssignable } from 'tsd';
 
-import { createDirectLine } from '../../src/boot/exports/full';
+import { createDirectLine } from '../../src/boot/actual/full';
 
 type CreateDirectLineInit = Parameters<typeof createDirectLine>[0];
