+++ packages/bundle/test-d/fail-once/invalid-prop-type-dir.test-d.tsx
@@ -3,7 +3,7 @@ import '../setup';
 import { type ComponentType } from 'react';
 import { expectNotAssignable } from 'tsd';
 
-import ReactWebChat from '../../src/boot/exports/full';
+import ReactWebChat from '../../src/boot/actual/full';
 
 type PropsOf<T> = T extends ComponentType<infer P> ? P : never;
