+++ packages/bundle/test-d/fail-once/invalid-style-options-suggested-actions-stacked-overflow.test-d.ts
@@ -2,7 +2,7 @@ import '../setup';
 
 import { expectNotAssignable } from 'tsd';
 
-import { createStyleSet } from '../../src/boot/exports/full';
+import { createStyleSet } from '../../src/boot/actual/full';
 
 type CreateStyleSetInit = Parameters<typeof createStyleSet>[0];
