+++ packages/bundle/test-d/fail-once/no-children-for-react-web-chat-in-minimal.test-d.tsx
@@ -4,7 +4,7 @@ import { expectNotAssignable } from 'tsd';
 
 import React, { type ComponentType } from 'react';
 
-import ReactWebChat from '../../src/boot/exports/full';
+import ReactWebChat from '../../src/boot/actual/full';
 
 type PropsOf<T> = T extends ComponentType<infer P> ? P : never;
