+++ packages/bundle/test-d/fail-once/no-full-bundle-style-options-in-minimal.test-d.ts
@@ -2,7 +2,7 @@ import '../setup';
 
 import { expectNotAssignable } from 'tsd';
 
-import { createStyleSet } from '../../src/boot/exports/minimal';
+import { createStyleSet } from '../../src/boot/actual/minimal';
 
 type CreateStyleSetInit = Parameters<typeof createStyleSet>[0];
