+++ packages/bundle/test-d/pass/correct-type-dir.test-d.tsx
@@ -2,7 +2,7 @@ import '../setup';
 
 import React from 'react';
 
-import ReactWebChat from '../../src/boot/exports/full';
+import ReactWebChat from '../../src/boot/actual/full';
 
 // eslint-disable-next-line react/jsx-indent
 <ReactWebChat dir="ltr" directLine={undefined} />;
