+++ packages/bundle/test-d/pass/import-create-direct-line.test-d.tsx
@@ -1,5 +1,5 @@
 import '../setup';
 
-import { createDirectLine } from '../../src/boot/exports/full';
+import { createDirectLine } from '../../src/boot/actual/full';
 
 createDirectLine({ token: 'faketoken' });
