+++ packages/bundle/test-d/pass/render-with-minimal-bundle.test-d.tsx
@@ -3,7 +3,7 @@ import '../setup';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ReactWebChat, { createDirectLine } from '../../src/boot/exports/minimal';
+import ReactWebChat, { createDirectLine } from '../../src/boot/actual/minimal';
 
 const directLine = createDirectLine({ token: '...' });
 const styleOptions = { accent: 'black' };
