+++ packages/bundle/test-d/pass/render-with-style-options.test-d.tsx
@@ -3,7 +3,7 @@ import '../setup';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ReactWebChat, { createDirectLine } from '../../src/boot/exports/full';
+import ReactWebChat, { createDirectLine } from '../../src/boot/actual/full';
 
 const directLine = createDirectLine({ token: '...' });
 const styleOptions = { accent: 'black', cardEmphasisBackgroundColor: 'orange' };
