+++ packages/bundle/test-d/pass/render.test-d.tsx
@@ -3,7 +3,7 @@ import '../setup';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ReactWebChat, { createDirectLine } from '../../src/boot/exports/full';
+import ReactWebChat, { createDirectLine } from '../../src/boot/actual/full';
 
 const directLine = createDirectLine({ token: '...' });
