+++ packages/bundle/test-d/pass/use-style-options-with-full-bundle.test-d.ts
@@ -1,6 +1,6 @@
 import '../setup';
 
-import { hooks } from '../../src/boot/exports/full';
+import { hooks } from '../../src/boot/actual/full';
 
 const [styleOptions] = hooks.useStyleOptions();
