+++ packages/bundle/test-d/pass/valid-style-options-suggested-actions-stacked-overflow.test-d.ts
@@ -1,6 +1,6 @@
 import '../setup';
 
-import { createStyleSet } from '../../src/boot/exports/full';
+import { createStyleSet } from '../../src/boot/actual/full';
 
 // Related to #4081.
 createStyleSet({ suggestedActionsStackedOverflow: 'auto' });
